+++ .gitignore
@@ -18,4 +18,4 @@
 .env
 /tmp
 
-/gobl.cfdi
+/gobl.mx.cfdi
