+++ cfdi.go
@@ -7,10 +7,10 @@ import (
 	"fmt"
 
 	"github.com/invopop/gobl"
-	"github.com/invopop/gobl.cfdi/addendas"
-	"github.com/invopop/gobl.cfdi/addon"
-	"github.com/invopop/gobl.cfdi/internal"
-	"github.com/invopop/gobl.cfdi/internal/format"
+	"github.com/invopop/gobl.mx.cfdi/addendas"
+	"github.com/invopop/gobl.mx.cfdi/addon"
+	"github.com/invopop/gobl.mx.cfdi/internal"
+	"github.com/invopop/gobl.mx.cfdi/internal/format"
 	"github.com/invopop/gobl/bill"
 	"github.com/invopop/gobl/cal"
 	"github.com/invopop/gobl/currency"
