+++ cmd/gobl.mx.cfdi/convert.go
@@ -6,7 +6,7 @@ import (
 	"io"
 
 	"github.com/invopop/gobl"
-	cfdi "github.com/invopop/gobl.cfdi"
+	cfdi "github.com/invopop/gobl.mx.cfdi"
 	"github.com/spf13/cobra"
 )
 
@@ -32,7 +32,7 @@ func (c *convertOpts) runE(cmd *cobra.Command, args []string) error {
 	// ctx := commandContext(cmd)
 
 	if len(args) == 0 || len(args) > 2 {
-		return fmt.Errorf("expected one or two arguments, the command usage is `gobl.cfdi convert <infile> [outfile]`")
+		return fmt.Errorf("expected one or two arguments, the command usage is `gobl.mx.cfdi convert <infile> [outfile]`")
 	}
 
 	input, err := openInput(cmd, args)
