+++ internal/lines.go
@@ -1,7 +1,7 @@
 package internal
 
 import (
-	"github.com/invopop/gobl.cfdi/addon"
+	"github.com/invopop/gobl.mx.cfdi/addon"
 	"github.com/invopop/gobl/bill"
 	"github.com/invopop/gobl/cbc"
 )
