+++ stamps_test.go
@@ -3,9 +3,9 @@ package cfdi_test
 import (
 	"testing"
 
-	cfdi "github.com/invopop/gobl.cfdi"
-	"github.com/invopop/gobl.cfdi/addon"
-	"github.com/invopop/gobl.cfdi/test"
+	cfdi "github.com/invopop/gobl.mx.cfdi"
+	"github.com/invopop/gobl.mx.cfdi/addon"
+	"github.com/invopop/gobl.mx.cfdi/test"
 	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/gobl/regimes/mx"
 	"github.com/stretchr/testify/assert"
